# api/swagger.yml
...
consumes:
- application/json
produces:
- application/json
swagger: "2.0"
info:
  description: API documentation
  title: github.com/ansiegl/Pok-Nest.git
  version: 0.1.0
paths:
...
  /api/v1/collection/pokemon/{pokemonId}:
    put:
      security:
      - Bearer: []
      description: edit a pokemon in user's collection
      tags:
      - collection
      summary: edit pokemon details in collection
      operationId: putEditPokemonInCollection
      parameters:
      - type: string
        description: pokemon's ID to edit
        name: pokemonId
        in: path
        required: true
      - name: caught
        in: body
        required: true
        schema:
          $ref: '#/definitions/collection'
      responses:
        "200":
          description: pokemon updated in collection successfully
        "400":
          description: invalid input data
        "404":
          description: pokemon or collection not found
    delete:
      security:
      - Bearer: []
      tags:
      - collection
      summary: delete pokemon from collection
      operationId: deletePokemonFromCollection
      parameters:
      - type: string
        description: pokemon's ID to delete
        name: pokemonId
        in: path
        required: true
      responses:
        "200":
          description: pokemon removed from collection successfully
        "404":
          description: pokemon or collection not found
...
  /api/v1/pokemon/{pokemonId}:
    post:
      security:
      - Bearer: []
      tags:
      - collection
      summary: add pokemon to collection
      operationId: postAddPokemonToCollection
      parameters:
      - type: string
        description: pokemon's ID to add to the collection
        name: pokemonId
        in: path
        required: true
      - name: caught
        in: body
        schema:
          $ref: '#/definitions/collection'
      responses:
        "201":
          description: pokemon added to collection successfully
        "400":
          description: invalid input data (e.g., invalid date or nickname)
        "404":
          description: pokemon or user not found
        "409":
          description: pokemon is already in the collection
...
definitions:
  collection:
    type: object
    required:
    - pokemon
    properties:
      pokemon:
        $ref: '#/definitions/pokemon'
...
  pokemon:
    type: object
    required:
    - pokemon_id
    - name
    - type1
    properties:
      generation:
        description: generation of the pokemon
        type: integer
        example: 1
      legendary:
        description: is the pokemon legendary?
        type: boolean
        example: false
      name:
        description: name of the pokemon
        type: string
        example: Pikachu
      pokemon_id:
        description: ID of the pokemon
        type: string
        format: uuid4
        example: 891d37d3-c74f-493e-aea8-af73efd92016
      type1:
        description: first type of the pokemon
        type: string
        example: Electric
      type2:
        description: second type of the pokemon
        type: string
        example: Flying
...
securityDefinitions:
  Bearer:
    description: |-
      Access token for application access, **must** include "Bearer " prefix.
      Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
    type: apiKey
    name: Authorization
    in: header
    x-keyPrefix: 'Bearer '
  Management:
    description: Management secret, used for monitoring and infrastructure related
      calls
    type: apiKey
    name: mgmt-secret
    in: query