# internal/types/collection.go
// Code generated by go-swagger; DO NOT EDIT.

package types

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Collection collection
//
// swagger:model collection
type Collection struct {

	// pokemon
	// Required: true
	Pokemon *Pokemon `json:"pokemon"`
}

// Validate validates this collection
func (m *Collection) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validatePokemon(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Collection) validatePokemon(formats strfmt.Registry) error {

	if err := validate.Required("pokemon", "body", m.Pokemon); err != nil {
		return err
	}

	if m.Pokemon != nil {
		if err := m.Pokemon.Validate(formats); err != nil {
			if ve, ok := err.(*errors.Validation); ok {
				return ve.ValidateName("pokemon")
			} else if ce, ok := err.(*errors.CompositeError); ok {
				return ce.ValidateName("pokemon")
			}
			return err
		}
	}

	return nil
}

// ContextValidate validate this collection based on the context it is used
func (m *Collection) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	var res []error

	if err := m.contextValidatePokemon(ctx, formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Collection) contextValidatePokemon(ctx context.Context, formats strfmt.Registry) error {

	if m.Pokemon != nil {
		if err := m.Pokemon.ContextValidate(ctx, formats); err != nil {
			if ve, ok := err.(*errors.Validation); ok {
				return ve.ValidateName("pokemon")
			} else if ce, ok := err.(*errors.CompositeError); ok {
				return ce.ValidateName("pokemon")
			}
			return err
		}
	}

	return nil
}

// MarshalBinary interface implementation
func (m *Collection) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Collection) UnmarshalBinary(b []byte) error {
	var res Collection
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}